refactor: align chatwoot folder structure with project conventions
Move the chatwoot app folder to the standard components/lib split used
elsewhere in the codebase:

  chatwoot/ChatwootWidget.tsx          -> components/chatwoot-widget.tsx
  chatwoot/actions.ts                  -> lib/actions.ts
  chatwoot/is-onboarding-pathname.ts   -> lib/utils.ts
  chatwoot/is-onboarding-pathname.test.ts -> lib/utils.test.ts

The .tsx file is also renamed to kebab-case to match neighbouring
components folders. Internal and external imports updated; tests pass.
